refactor(server): split server.go into 6 responsibility files
Splits the 3401-line server.go into:
- server_helpers.go (util/pointer helpers)
- server_metadata.go (metadata state, book enrichment)
- server_search.go (search index management)
- server_lifecycle.go (startup, routes, stale ops)
- server_middleware.go (CORS, path guards, iTunes sync)
- server_title_helpers.go (title/series extraction)

Types, vars, consts, NewServer stay in server.go.
No logic changes. Structure audit STRUCT-7.
