+++ internal/server/server_helpers.go
@@ -1,0 +1,133 @@
+// file: internal/server/server_helpers.go
+// version: 1.0.0
+// guid: 8a40b808-2bf2-4a35-893c-ad5e3351dbae
+// last-edited: 2026-05-01
+
+package server
+
+import (
+	"os"
+	"path/filepath"
+	"strings"
+	"time"
+
+	"github.com/jdfalk/audiobook-organizer/internal/database"
+)
+
+func SetVersion(v string) {
+	appVersion = v
+}
+
+// resetLibrarySizeCache resets the library size cache (for testing)
+func resetLibrarySizeCache() {
+	cacheLock.Lock()
+	defer cacheLock.Unlock()
+	cachedLibrarySize = 0
+	cachedImportSize = 0
+	cachedSizeComputedAt = time.Time{}
+}
+
+// Helper functions for pointer conversions
+func stringPtr(s string) *string {
+	return &s
+}
+
+func intPtrHelper(i int) *int {
+	return &i
+}
+
+func boolPtr(b bool) *bool {
+	return &b
+}
+
+func ptrStr(p *string) string {
+	if p == nil {
+		return ""
+	}
+	return *p
+}
+
+func stringVal(p *string) any {
+	if p == nil {
+		return nil
+	}
+	return *p
+}
+
+func intVal(p *int) any {
+	if p == nil {
+		return nil
+	}
+	return *p
+}
+
+func nonEmpty(s string) any {
+	if s == "" {
+		return nil
+	}
+	return s
+}
+
+func calculateLibrarySizes(rootDir string, importFolders []database.ImportPath) (librarySize, importSize int64) {
+	cacheLock.RLock()
+	if time.Since(cachedSizeComputedAt) < librarySizeCacheTTL {
+		librarySize = cachedLibrarySize
+		importSize = cachedImportSize
+		cacheLock.RUnlock()
+		// cached sizes used
+		return
+	}
+	cacheLock.RUnlock()
+
+	// Cache expired, recalculate
+	cacheLock.Lock()
+	defer cacheLock.Unlock()
+
+	// Double-check in case another goroutine just updated
+	if time.Since(cachedSizeComputedAt) < librarySizeCacheTTL {
+		return cachedLibrarySize, cachedImportSize
+	}
+
+	// Recalculating library sizes (cache expired)
+
+	// Calculate library size
+	librarySize = 0
+	if rootDir != "" {
+		if info, err := os.Stat(rootDir); err == nil && info.IsDir() {
+			filepath.Walk(rootDir, func(path string, info os.FileInfo, err error) error {
+				if err == nil && !info.IsDir() {
+					librarySize += filePhysicalSize(info)
+				}
+				return nil
+			})
+		}
+	}
+
+	// Calculate import path sizes independently (not by subtraction)
+	importSize = 0
+	for _, folder := range importFolders {
+		if !folder.Enabled {
+			continue
+		}
+		if info, err := os.Stat(folder.Path); err == nil && info.IsDir() {
+			filepath.Walk(folder.Path, func(path string, info os.FileInfo, err error) error {
+				if err == nil && !info.IsDir() {
+					// Skip files that are under rootDir to avoid double counting
+					if rootDir != "" && strings.HasPrefix(path, rootDir) {
+						return nil
+					}
+					importSize += filePhysicalSize(info)
+				}
+				return nil
+			})
+		}
+	}
+
+	// Update cache
+	cachedLibrarySize = librarySize
+	cachedImportSize = importSize
+	cachedSizeComputedAt = time.Now()
+
+	// sizes recalculated
+	return
+}
