+++ internal/server/server_metadata.go
@@ -1,0 +1,558 @@
+// file: internal/server/server_metadata.go
+// version: 1.0.0
+// guid: 588350bc-83db-47ed-9590-2b6513aadcda
+// last-edited: 2026-05-01
+
+package server
+
+import (
+	"encoding/json"
+	"fmt"
+	"log"
+	"os"
+	"strings"
+	"time"
+
+	"github.com/jdfalk/audiobook-organizer/internal/activity"
+	"github.com/jdfalk/audiobook-organizer/internal/database"
+	"github.com/jdfalk/audiobook-organizer/internal/metadata"
+	"github.com/jdfalk/audiobook-organizer/internal/scanner"
+)
+
+func metadataStateKey(bookID string) string {
+	return fmt.Sprintf("metadata_state_%s", bookID)
+}
+
+func decodeMetadataValue(raw *string) any {
+	if raw == nil || *raw == "" {
+		return nil
+	}
+	var value any
+	if err := json.Unmarshal([]byte(*raw), &value); err != nil {
+		return *raw
+	}
+	return value
+}
+
+func encodeMetadataValue(value any) (*string, error) {
+	if value == nil {
+		return nil, nil
+	}
+	data, err := json.Marshal(value)
+	if err != nil {
+		return nil, err
+	}
+	encoded := string(data)
+	return &encoded, nil
+}
+
+func loadLegacyMetadataState(bookID string) (map[string]metadataFieldState, error) {
+	state := map[string]metadataFieldState{}
+
+	pref, err := database.GetGlobalStore().GetUserPreference(metadataStateKey(bookID))
+	if err != nil {
+		return state, err
+	}
+	if pref == nil || pref.Value == nil || *pref.Value == "" {
+		return state, nil
+	}
+
+	if err := json.Unmarshal([]byte(*pref.Value), &state); err != nil {
+		return state, fmt.Errorf("failed to parse metadata state: %w", err)
+	}
+	return state, nil
+}
+
+func loadMetadataState(bookID string) (map[string]metadataFieldState, error) {
+	state := map[string]metadataFieldState{}
+	if database.GetGlobalStore() == nil {
+		return state, fmt.Errorf("database not initialized")
+	}
+
+	stored, err := database.GetGlobalStore().GetMetadataFieldStates(bookID)
+	if err != nil {
+		return state, err
+	}
+	for _, entry := range stored {
+		state[entry.Field] = metadataFieldState{
+			FetchedValue:   decodeMetadataValue(entry.FetchedValue),
+			OverrideValue:  decodeMetadataValue(entry.OverrideValue),
+			OverrideLocked: entry.OverrideLocked,
+			UpdatedAt:      entry.UpdatedAt,
+		}
+	}
+	if len(state) > 0 {
+		return state, nil
+	}
+
+	legacy, err := loadLegacyMetadataState(bookID)
+	if err != nil {
+		return state, err
+	}
+	if len(legacy) == 0 {
+		return state, nil
+	}
+
+	if err := saveMetadataState(bookID, legacy); err != nil {
+		log.Printf("[WARN] failed to migrate legacy metadata state for %s: %v", bookID, err)
+	}
+	return legacy, nil
+}
+
+func saveMetadataState(bookID string, state map[string]metadataFieldState) error {
+	if database.GetGlobalStore() == nil {
+		return fmt.Errorf("database not initialized")
+	}
+
+	existing, err := database.GetGlobalStore().GetMetadataFieldStates(bookID)
+	if err != nil {
+		return err
+	}
+	existingFields := map[string]struct{}{}
+	for _, entry := range existing {
+		existingFields[entry.Field] = struct{}{}
+	}
+
+	now := time.Now()
+	for field, entry := range state {
+		fetched, err := encodeMetadataValue(entry.FetchedValue)
+		if err != nil {
+			return fmt.Errorf("failed to encode fetched metadata for %s: %w", field, err)
+		}
+		override, err := encodeMetadataValue(entry.OverrideValue)
+		if err != nil {
+			return fmt.Errorf("failed to encode override metadata for %s: %w", field, err)
+		}
+		if entry.UpdatedAt.IsZero() {
+			entry.UpdatedAt = now
+		}
+
+		dbState := database.MetadataFieldState{
+			BookID:         bookID,
+			Field:          field,
+			FetchedValue:   fetched,
+			OverrideValue:  override,
+			OverrideLocked: entry.OverrideLocked,
+			UpdatedAt:      entry.UpdatedAt,
+		}
+
+		if err := database.GetGlobalStore().UpsertMetadataFieldState(&dbState); err != nil {
+			return fmt.Errorf("failed to persist metadata state for %s: %w", field, err)
+		}
+		delete(existingFields, field)
+	}
+
+	for field := range existingFields {
+		if err := database.GetGlobalStore().DeleteMetadataFieldState(bookID, field); err != nil {
+			return fmt.Errorf("failed to clean up metadata state for %s: %w", field, err)
+		}
+	}
+
+	return nil
+}
+
+func decodeRawValue(raw json.RawMessage) any {
+	if raw == nil {
+		return nil
+	}
+	var value any
+	if err := json.Unmarshal(raw, &value); err != nil {
+		return string(raw)
+	}
+	return value
+}
+
+func updateFetchedMetadataState(bookID string, values map[string]any) error {
+	state, err := loadMetadataState(bookID)
+	if err != nil {
+		return err
+	}
+	if state == nil {
+		state = map[string]metadataFieldState{}
+	}
+	for field, value := range values {
+		entry := state[field]
+		entry.FetchedValue = value
+		entry.UpdatedAt = time.Now()
+		state[field] = entry
+	}
+	return saveMetadataState(bookID, state)
+}
+
+func resolveAuthorAndSeriesNames(book *database.Book) (string, string) {
+	authorName := ""
+	if book.Author != nil {
+		authorName = book.Author.Name
+	} else if book.AuthorID != nil {
+		if author, err := database.GetGlobalStore().GetAuthorByID(*book.AuthorID); err == nil && author != nil {
+			authorName = author.Name
+		}
+	}
+
+	seriesName := ""
+	if book.Series != nil {
+		seriesName = book.Series.Name
+	} else if book.SeriesID != nil {
+		if series, err := database.GetGlobalStore().GetSeriesByID(*book.SeriesID); err == nil && series != nil {
+			seriesName = series.Name
+		}
+	}
+
+	return authorName, seriesName
+}
+
+// batchFetchBookAuthorsAndNarrators pre-fetches author and narrator join table
+// entries plus their full details for all given books. Returns maps keyed by
+// book ID for join entries, plus maps keyed by author/narrator ID for details.
+// Nil maps are returned if the global store is not available.
+func batchFetchBookAuthorsAndNarrators(bookIDs []string) (map[string][]database.BookAuthor, map[int]*database.Author, map[string][]database.BookNarrator, map[int]*database.Narrator) {
+	if len(bookIDs) == 0 || database.GetGlobalStore() == nil {
+		return nil, nil, nil, nil
+	}
+
+	store := database.GetGlobalStore()
+
+	// Collect all book authors and extract author IDs
+	bookAuthorsMap := make(map[string][]database.BookAuthor)
+	authorIDs := make(map[int]bool)
+	for _, bookID := range bookIDs {
+		if bas, err := store.GetBookAuthors(bookID); err == nil {
+			bookAuthorsMap[bookID] = bas
+			for _, ba := range bas {
+				authorIDs[ba.AuthorID] = true
+			}
+		}
+	}
+
+	// Fetch all authors in bulk
+	authorsByID := make(map[int]*database.Author)
+	for authorID := range authorIDs {
+		if author, err := store.GetAuthorByID(authorID); err == nil && author != nil {
+			authorsByID[authorID] = author
+		}
+	}
+
+	// Collect all book narrators and extract narrator IDs
+	bookNarratorsMap := make(map[string][]database.BookNarrator)
+	narratorIDs := make(map[int]bool)
+	for _, bookID := range bookIDs {
+		if bns, err := store.GetBookNarrators(bookID); err == nil {
+			bookNarratorsMap[bookID] = bns
+			for _, bn := range bns {
+				narratorIDs[bn.NarratorID] = true
+			}
+		}
+	}
+
+	// Fetch all narrators in bulk
+	narratorsByID := make(map[int]*database.Narrator)
+	for narratorID := range narratorIDs {
+		if narrator, err := store.GetNarratorByID(narratorID); err == nil && narrator != nil {
+			narratorsByID[narratorID] = narrator
+		}
+	}
+
+	return bookAuthorsMap, authorsByID, bookNarratorsMap, narratorsByID
+}
+
+// enrichBookForResponseSingle enriches a single book by pre-fetching its
+// author and narrator data. Convenience wrapper for single-book endpoints.
+func enrichBookForResponseSingle(book *database.Book) enrichedBookResponse {
+	bookAuthorsMap, authorsByID, bookNarratorsMap, narratorsByID := batchFetchBookAuthorsAndNarrators([]string{book.ID})
+	return enrichBookForResponse(book, bookAuthorsMap, authorsByID, bookNarratorsMap, narratorsByID)
+}
+
+// enrichBookForResponse resolves author, series, and narrator names from join
+// tables so the JSON response contains all the fields the frontend expects.
+// Pre-fetched maps of authors and narrators (by book ID) are used instead of
+// per-book DB calls to eliminate N+1 queries.
+func enrichBookForResponse(book *database.Book, bookAuthorsMap map[string][]database.BookAuthor, authorsByID map[int]*database.Author, bookNarratorsMap map[string][]database.BookNarrator, narratorsByID map[int]*database.Narrator) enrichedBookResponse {
+	authorName, seriesName := resolveAuthorAndSeriesNames(book)
+	resp := enrichedBookResponse{Book: book}
+	if authorName != "" {
+		resp.AuthorName = &authorName
+	}
+	if seriesName != "" {
+		resp.SeriesName = &seriesName
+	}
+
+	// Check if the book's file exists on disk (single-file books only).
+	if book.FilePath != "" {
+		_, statErr := os.Stat(book.FilePath)
+		exists := statErr == nil
+		resp.FileExists = &exists
+	}
+
+	if bookAuthorsMap != nil && authorsByID != nil {
+		if bookAuthors, ok := bookAuthorsMap[book.ID]; ok && len(bookAuthors) > 0 {
+			for _, ba := range bookAuthors {
+				if author, ok := authorsByID[ba.AuthorID]; ok && author != nil {
+					resp.Authors = append(resp.Authors, authorEntry{
+						ID: author.ID, Name: author.Name, Role: ba.Role, Position: ba.Position,
+					})
+				}
+			}
+			if resp.AuthorName == nil && len(resp.Authors) > 0 {
+				names := make([]string, len(resp.Authors))
+				for i, a := range resp.Authors {
+					names[i] = a.Name
+				}
+				combined := strings.Join(names, " & ")
+				resp.AuthorName = &combined
+			}
+		}
+	}
+
+	if bookNarratorsMap != nil && narratorsByID != nil {
+		if bookNarrators, ok := bookNarratorsMap[book.ID]; ok && len(bookNarrators) > 0 {
+			for _, bn := range bookNarrators {
+				if narrator, ok := narratorsByID[bn.NarratorID]; ok && narrator != nil {
+					resp.Narrators = append(resp.Narrators, narratorEntry{
+						ID: narrator.ID, Name: narrator.Name, Role: bn.Role, Position: bn.Position,
+					})
+				}
+			}
+			if (book.Narrator == nil || *book.Narrator == "") && len(resp.Narrators) > 0 {
+				names := make([]string, len(resp.Narrators))
+				for i, n := range resp.Narrators {
+					names[i] = n.Name
+				}
+				combined := strings.Join(names, " & ")
+				book.Narrator = &combined
+			}
+		}
+	}
+
+	// Populate metadata_source_hash_duplicate_count if this book has a hash.
+	// This lets the BookDetail UI warn about possible duplicates without an
+	// extra round-trip.
+	if book.MetadataSourceHash != nil && *book.MetadataSourceHash != "" && database.GetGlobalStore() != nil {
+		if matches, err := database.GetGlobalStore().GetBooksByMetadataSourceHash(*book.MetadataSourceHash); err == nil {
+			count := 0
+			for _, m := range matches {
+				if m.ID != book.ID {
+					count++
+				}
+			}
+			if count > 0 {
+				resp.MetadataSourceHashDuplicateCount = &count
+			}
+		}
+	}
+
+	return resp
+}
+
+func buildComparisonValuesFromMetadata(comparisonMeta *metadata.Metadata) map[string]any {
+	if comparisonMeta == nil {
+		return nil
+	}
+
+	compMap := map[string]any{
+		"title":           nonEmpty(comparisonMeta.Title),
+		"author_name":     nonEmpty(comparisonMeta.Artist),
+		"narrator":        nonEmpty(comparisonMeta.Narrator),
+		"series_name":     nonEmpty(comparisonMeta.Series),
+		"publisher":       nonEmpty(comparisonMeta.Publisher),
+		"language":        nonEmpty(comparisonMeta.Language),
+		"isbn10":          nonEmpty(comparisonMeta.ISBN10),
+		"isbn13":          nonEmpty(comparisonMeta.ISBN13),
+		"genre":           nonEmpty(comparisonMeta.Genre),
+		"album":           nonEmpty(comparisonMeta.Album),
+		"asin":            nonEmpty(comparisonMeta.ASIN),
+		"edition":         nonEmpty(comparisonMeta.Edition),
+		"print_year":      nonEmpty(comparisonMeta.PrintYear),
+		"description":     nonEmpty(comparisonMeta.Comments),
+		"book_id":         nonEmpty(comparisonMeta.BookOrganizerID),
+		"open_library_id": nonEmpty(comparisonMeta.OpenLibraryID),
+		"hardcover_id":    nonEmpty(comparisonMeta.HardcoverID),
+		"google_books_id": nonEmpty(comparisonMeta.GoogleBooksID),
+	}
+	if comparisonMeta.Year > 0 {
+		compMap["audiobook_release_year"] = comparisonMeta.Year
+	}
+	if comparisonMeta.SeriesIndex > 0 {
+		compMap["series_index"] = comparisonMeta.SeriesIndex
+	}
+	return compMap
+}
+
+func buildComparisonValuesFromBook(book *database.Book, authorName, seriesName string) map[string]any {
+	if book == nil {
+		return nil
+	}
+
+	compMap := map[string]any{
+		"title":           nonEmpty(book.Title),
+		"author_name":     nonEmpty(authorName),
+		"narrator":        nonEmpty(ptrStr(book.Narrator)),
+		"series_name":     nonEmpty(seriesName),
+		"publisher":       nonEmpty(ptrStr(book.Publisher)),
+		"language":        nonEmpty(ptrStr(book.Language)),
+		"isbn10":          nonEmpty(ptrStr(book.ISBN10)),
+		"isbn13":          nonEmpty(ptrStr(book.ISBN13)),
+		"genre":           nonEmpty(ptrStr(book.Genre)),
+		"album":           nonEmpty(book.Title),
+		"asin":            nonEmpty(ptrStr(book.ASIN)),
+		"edition":         nonEmpty(ptrStr(book.Edition)),
+		"description":     nonEmpty(ptrStr(book.Description)),
+		"book_id":         nonEmpty(book.ID),
+		"open_library_id": nonEmpty(ptrStr(book.OpenLibraryID)),
+		"hardcover_id":    nonEmpty(ptrStr(book.HardcoverID)),
+		"google_books_id": nonEmpty(ptrStr(book.GoogleBooksID)),
+	}
+	if book.AudiobookReleaseYear != nil && *book.AudiobookReleaseYear > 0 {
+		compMap["audiobook_release_year"] = *book.AudiobookReleaseYear
+	}
+	if book.SeriesSequence != nil && *book.SeriesSequence > 0 {
+		compMap["series_index"] = *book.SeriesSequence
+	}
+	if book.PrintYear != nil && *book.PrintYear > 0 {
+		compMap["print_year"] = *book.PrintYear
+	}
+	return compMap
+}
+
+// buildComparisonValuesFromActivityLog reconstructs a "before" tag snapshot by
+// querying the activity log for metadata_apply entries for the given book
+// recorded within a ±5 second window of ts. For each field found, the
+// old_value (i.e. the value BEFORE that operation) is used as the comparison
+// value. This is the fallback when GetBookAtVersion is unavailable (SQLite) or
+// when the exact version key is not present in PebbleDB.
+func buildComparisonValuesFromActivityLog(as *activity.Service, bookID string, ts time.Time) map[string]any {
+	window := 5 * time.Second
+	since := ts.Add(-window)
+	until := ts.Add(window)
+
+	entries, _, err := as.Query(database.ActivityFilter{
+		BookID: bookID,
+		Type:   "metadata_apply",
+		Since:  &since,
+		Until:  &until,
+		Limit:  200,
+	})
+	if err != nil || len(entries) == 0 {
+		return nil
+	}
+
+	compMap := map[string]any{}
+	for _, e := range entries {
+		if e.Details == nil {
+			continue
+		}
+		field, _ := e.Details["field"].(string)
+		if field == "" {
+			continue
+		}
+		// old_value is the state BEFORE this operation — that's what we want
+		// to show as the "snapshot" comparison row.
+		if oldVal, ok := e.Details["old_value"]; ok && oldVal != nil {
+			if s, ok := oldVal.(string); ok && s != "" {
+				compMap[field] = s
+			} else if oldVal != nil {
+				compMap[field] = oldVal
+			}
+		}
+	}
+	if len(compMap) == 0 {
+		return nil
+	}
+	return compMap
+}
+
+func buildMetadataProvenance(book *database.Book, state map[string]metadataFieldState, meta metadata.Metadata, authorName, seriesName string, comparisonValues map[string]any) map[string]database.MetadataProvenanceEntry {
+	if state == nil {
+		state = map[string]metadataFieldState{}
+	}
+
+	provenance := map[string]database.MetadataProvenanceEntry{}
+
+	addEntry := func(field string, fileValue any, storedValue any) {
+		entryState := state[field]
+		effectiveSource := ""
+		var effectiveValue any
+		switch {
+		case entryState.OverrideValue != nil:
+			effectiveSource = "override"
+			effectiveValue = entryState.OverrideValue
+		case storedValue != nil:
+			effectiveSource = "stored"
+			effectiveValue = storedValue
+		case entryState.FetchedValue != nil:
+			effectiveSource = "fetched"
+			effectiveValue = entryState.FetchedValue
+		case fileValue != nil:
+			effectiveSource = "file"
+			effectiveValue = fileValue
+		}
+
+		var updatedAt *time.Time
+		if !entryState.UpdatedAt.IsZero() {
+			ts := entryState.UpdatedAt.UTC()
+			updatedAt = &ts
+		}
+
+		entry := database.MetadataProvenanceEntry{
+			FileValue:       fileValue,
+			FetchedValue:    entryState.FetchedValue,
+			StoredValue:     storedValue,
+			OverrideValue:   entryState.OverrideValue,
+			OverrideLocked:  entryState.OverrideLocked,
+			EffectiveValue:  effectiveValue,
+			EffectiveSource: effectiveSource,
+			UpdatedAt:       updatedAt,
+		}
+
+		if comparisonValues != nil {
+			if cv, ok := comparisonValues[field]; ok {
+				entry.ComparisonValue = cv
+			}
+		}
+
+		provenance[field] = entry
+	}
+
+	addEntry("title", meta.Title, book.Title)
+	addEntry("author_name", meta.Artist, authorName)
+	addEntry("narrator", meta.Narrator, stringVal(book.Narrator))
+	addEntry("series_name", meta.Series, seriesName)
+	addEntry("publisher", meta.Publisher, stringVal(book.Publisher))
+	addEntry("language", meta.Language, stringVal(book.Language))
+	addEntry("audiobook_release_year", meta.Year, intVal(book.AudiobookReleaseYear))
+	addEntry("isbn10", meta.ISBN10, stringVal(book.ISBN10))
+	addEntry("isbn13", meta.ISBN13, stringVal(book.ISBN13))
+	addEntry("genre", meta.Genre, stringVal(book.Genre))
+	addEntry("album", meta.Album, book.Title)
+	addEntry("asin", nonEmpty(meta.ASIN), stringVal(book.ASIN))
+	var seriesIdx any
+	if meta.SeriesIndex > 0 {
+		seriesIdx = meta.SeriesIndex
+	}
+	addEntry("series_index", seriesIdx, intVal(book.SeriesSequence))
+	addEntry("print_year", nonEmpty(meta.PrintYear), intVal(book.PrintYear))
+	addEntry("edition", nonEmpty(meta.Edition), stringVal(book.Edition))
+	addEntry("description", nonEmpty(meta.Comments), stringVal(book.Description))
+	addEntry("book_id", nonEmpty(meta.BookOrganizerID), book.ID)
+	addEntry("open_library_id", nonEmpty(meta.OpenLibraryID), stringVal(book.OpenLibraryID))
+	addEntry("hardcover_id", nonEmpty(meta.HardcoverID), stringVal(book.HardcoverID))
+	addEntry("google_books_id", nonEmpty(meta.GoogleBooksID), stringVal(book.GoogleBooksID))
+
+	return provenance
+}
+
+func applyOrganizedFileMetadata(book *database.Book, newPath string) {
+	hash, err := scanner.ComputeFileHash(newPath)
+	if err != nil {
+		log.Printf("[WARN] failed to compute organized hash for %s: %v", newPath, err)
+	} else if hash != "" {
+		book.FileHash = stringPtr(hash)
+		book.OrganizedFileHash = stringPtr(hash)
+		if book.OriginalFileHash == nil {
+			book.OriginalFileHash = stringPtr(hash)
+		}
+	}
+	if info, err := os.Stat(newPath); err == nil {
+		size := info.Size()
+		book.FileSize = &size
+	}
+}
